Moved voting into it's own component
Previously, voting with the slider made each slider change cause updates ont he component. Having that happen int he PostSummary and PostDetails pages was causing excessive upadtes for nothing, and degraded perfomance.

Moving the functionality into it's own component made updates only run on the voting component, not on the whole post summary or full details.
